+++ src/app/projects/assets/packages.json
@@ -1,0 +1,23 @@
+[
+  {
+    "name": "fmdapi-node-weaver",
+    "day": 0,
+    "week": 3,
+    "year": 70,
+    "total": 70
+  },
+  {
+    "name": "react-canvas-editor",
+    "day": 1,
+    "week": 22,
+    "year": 947,
+    "total": 947
+  },
+  {
+    "name": "canvas-editor",
+    "day": 0,
+    "week": 7,
+    "year": 1063,
+    "total": 1063
+  }
+]
